+++ app/src/main/java/mypoli/android/common/AppState.kt
@@ -7,10 +7,12 @@ import mypoli.android.common.redux.Reducer
 import mypoli.android.common.redux.State
 import mypoli.android.pet.store.PetStoreReducer
 import mypoli.android.pet.store.PetStoreState
-import mypoli.android.quest.agenda.AgendaReducer
-import mypoli.android.quest.agenda.AgendaState
-import mypoli.android.quest.calendar.CalendarReducer
-import mypoli.android.quest.calendar.CalendarState
+import mypoli.android.quest.schedule.ScheduleReducer
+import mypoli.android.quest.schedule.ScheduleState
+import mypoli.android.quest.schedule.agenda.AgendaReducer
+import mypoli.android.quest.schedule.agenda.AgendaState
+import mypoli.android.quest.schedule.calendar.CalendarReducer
+import mypoli.android.quest.schedule.calendar.CalendarState
 
 /**
  * Created by Venelin Valkov <venelin@ipoli.io>
@@ -25,6 +27,7 @@ sealed class LoadDataAction : Action {
 
 data class AppState(
     val appDataState: AppDataState,
+    val scheduleState: ScheduleState,
     val calendarState: CalendarState,
     val agendaState: AgendaState,
     val petStoreState: PetStoreState,
@@ -34,15 +37,17 @@ data class AppState(
 object AppReducer : Reducer<AppState, AppState> {
 
     override fun reduce(state: AppState, action: Action) = state.copy(
+        appDataState = AppDataReducer.reduce(state, action),
+        scheduleState = ScheduleReducer.reduce(state, action),
         calendarState = CalendarReducer.reduce(state, action),
         agendaState = AgendaReducer.reduce(state, action),
-        appDataState = AppDataReducer.reduce(state, action),
         petStoreState = PetStoreReducer.reduce(state, action)
     )
 
     override fun defaultState() =
         AppState(
             appDataState = AppDataReducer.defaultState(),
+            scheduleState = ScheduleReducer.defaultState(),
             calendarState = CalendarReducer.defaultState(),
             agendaState = AgendaReducer.defaultState(),
             petStoreState = PetStoreReducer.defaultState(),
