--- app/src/main/java/mypoli/android/quest/calendar/dayview/view/widget/CalendarEvent.kt
@@ -1,14 +1,0 @@
-package mypoli.android.quest.calendar.dayview.view.widget
-
-/**
- * Created by Venelin Valkov <venelin@mypoli.fun>
- * on 9/2/17.
- */
-interface CalendarEvent {
-
-    val id: String
-
-    val duration: Int
-
-    val startMinute: Int
-}
