+++ app/src/main/java/mypoli/android/quest/schedule/calendar/dayview/view/widget/CalendarEvent.kt
@@ -1,0 +1,14 @@
+package mypoli.android.quest.schedule.calendar.dayview.view.widget
+
+/**
+ * Created by Venelin Valkov <venelin@mypoli.fun>
+ * on 9/2/17.
+ */
+interface CalendarEvent {
+
+    val id: String
+
+    val duration: Int
+
+    val startMinute: Int
+}
