+++ app/src/main/res/layout/controller_calendar.xml
@@ -1,117 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
+<mypoli.android.quest.schedule.calendar.dayview.view.widget.LockableViewPager xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/pager"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:clipChildren="false"
-    android:orientation="vertical"
-    android:weightSum="1">
-
-    <RelativeLayout
-        android:id="@+id/datePickerContainer"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:clipChildren="false"
-        android:elevation="8dp">
-
-        <TextView
-            android:id="@+id/currentMonth"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_alignParentTop="true"
-            android:background="@color/md_white"
-            android:gravity="center"
-            android:paddingBottom="8dp"
-            android:paddingTop="8dp"
-            android:textAppearance="@style/TextAppearance.AppCompat.Title" />
-
-        <sun.bob.mcalendarview.views.WeekColumnView
-            android:id="@+id/weekColumnView"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_below="@+id/currentMonth"
-            android:paddingBottom="8dp"
-            android:paddingTop="8dp" />
-
-        <View
-            android:id="@+id/separator"
-            android:layout_width="match_parent"
-            android:layout_height="1px"
-            android:layout_below="@+id/weekColumnView"
-            android:background="#ccc" />
-
-        <sun.bob.mcalendarview.views.ExpCalendarView
-            android:id="@+id/datePicker"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_below="@+id/separator" />
-
-        <ImageView
-            android:id="@+id/expander"
-            android:layout_width="28dp"
-            android:layout_height="28dp"
-            android:layout_alignBottom="@+id/datePicker"
-            android:layout_centerHorizontal="true"
-            android:layout_marginBottom="-14dp"
-            android:background="@drawable/circle_white"
-            android:elevation="4dp"
-            android:scaleType="fitXY" />
-
-        <View
-            android:layout_width="match_parent"
-            android:layout_height="12dp"
-            android:layout_below="@+id/datePicker"
-            android:background="@android:color/transparent" />
-    </RelativeLayout>
-
-    <ProgressBar
-        android:id="@+id/levelProgress"
-        style="@style/Widget.AppCompat.ProgressBar.Horizontal"
-        android:layout_width="match_parent"
-        android:layout_height="6dp"
-        android:elevation="3dp"
-        android:progressDrawable="@drawable/experience_progress_bar" />
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:elevation="2dp"
-        android:id="@+id/contentContainer"
-        app:layout_behavior="@string/appbar_scrolling_view_behavior">
-
-        <mypoli.android.quest.calendar.dayview.view.widget.LockableViewPager
-            android:id="@+id/pager"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:elevation="2dp" />
-    </LinearLayout>
-
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/addQuest"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|end"
-        android:layout_marginBottom="16dp"
-        android:layout_marginEnd="16dp"
-        android:src="@drawable/ic_add_white_24dp"
-        app:elevation="6dp" />
-
-    <View
-        android:id="@+id/addContainerBackground"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@color/md_dark_text_26"
-        android:clickable="true"
-        android:elevation="3dp"
-        android:focusable="true"
-        android:visibility="gone" />
-
-    <com.bluelinelabs.conductor.ChangeHandlerFrameLayout
-        android:id="@+id/addContainer"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom"
-        android:elevation="3dp"
-        android:visibility="invisible" />
-
-</android.support.design.widget.CoordinatorLayout>
+    android:elevation="2dp" />
