--- app/src/main/res/menu/calendar_menu.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item
-        android:id="@+id/actionAgenda"
-        android:icon="@drawable/ic_format_list_bulleted_white_24dp"
-        android:title="Agenda"
-        app:showAsAction="always" />
-</menu>
