+++ app/src/main/res/menu/schedule_menu.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <item
+        android:id="@+id/actionViewMode"
+        android:icon="@drawable/ic_format_list_bulleted_white_24dp"
+        android:title="Agenda"
+        app:showAsAction="always" />
+</menu>
